feat: delivery/company/product service feign client edit

+++ company-service/src/main/java/on/logistics/companyservice/application/service/CompanyServiceImpl.java
@@ -1,5 +1,6 @@
 package on.logistics.companyservice.application.service;
 
+import jakarta.servlet.http.HttpServletRequest;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -15,6 +16,13 @@ import on.logistics.companyservice.domain.repository.CompanyRepository;
 import on.logistics.companyservice.exception.CompanyException;
 import on.logistics.companyservice.exception.CompanyExceptionCode;
 import on.logistics.companyservice.global.application.dtos.PageDto;
+import on.logistics.companyservice.global.domain.Passport;
+import on.logistics.companyservice.global.enums.AuthRole;
+import on.logistics.companyservice.global.utils.PassportUtil;
+import on.logistics.companyservice.infrastructure.clients.hub.HubServiceClient;
+import on.logistics.companyservice.infrastructure.clients.hub.feign.dtos.GetHubInfo;
+import on.logistics.companyservice.infrastructure.clients.hub.feign.dtos.GetHubManagerBooleanResponse;
+import on.logistics.companyservice.infrastructure.clients.hub.feign.dtos.HubManagerBooleanRequest;
 import on.logistics.companyservice.presentation.dtos.response.CreateCompanyResponse;
 import on.logistics.companyservice.presentation.dtos.response.GetCompanyResponse;
 import on.logistics.companyservice.presentation.dtos.response.SearchCompanyResponse;
@@ -33,19 +41,29 @@ import org.springframework.transaction.annotation.Transactional;
 public class CompanyServiceImpl implements CompanyService {
 
     private final CompanyRepository companyRepository;
+    private final HubServiceClient hubServiceClient;
+    private final PassportUtil passportUtil;
 
     @Override
     @Transactional
     public CreateCompanyResponse createCompany(CreateCompanyRequestDto requestDto) {
-        // todo : 임시 유저 아이디 발급 로직 수정 필요
-        UUID userId = UUID.randomUUID();
+        Passport passport = getPassport(requestDto.passportRequest());
+        notMasterAndNotHubManagerValid(passport);
 
-        companyRepository.findByUserId(userId).ifPresent(company -> {
+        companyRepository.findByUserId(passport.getUserId()).ifPresent(company -> {
             throw new CompanyException(CompanyExceptionCode.COMPANY_USER_ID_DUPLICATE);
         });
 
-        CreateCompanyDto createCompanyDto = CreateCompanyDto.from(userId, requestDto.companyName(),
-            requestDto.companyType(), requestDto.companyAddress());
+        GetHubInfo hubInfo = hubServiceClient.getHubInfo(requestDto.managedHubId());
+        if (hubInfo == null) {
+            throw new CompanyException(CompanyExceptionCode.COMPANY_HUB_NOT_FOUND);
+        }
+
+        validHubManagerHub(passport, hubInfo.id());
+
+        CreateCompanyDto createCompanyDto = CreateCompanyDto.from(requestDto.userId(),
+            requestDto.companyName(), requestDto.companyType(), requestDto.companyAddress(),
+            requestDto.managedHubId());
         Company company = Company.create(createCompanyDto);
         Company saved = companyRepository.save(company);
         return CreateCompanyResponse.of(saved.getId());
@@ -61,24 +79,42 @@ public class CompanyServiceImpl implements CompanyService {
     @Override
     public GetCompanyResponse getCompany(UUID id) {
         Company company = getOrElseThrow(id);
-        return GetCompanyResponse.of(company.getId(), company.getName().getValue(),
-            company.getType(), company.getManagedHubId(), company.getAddress().getValue());
+        return GetCompanyResponse.of(company.getId(), company.getUserId(),
+            company.getName().getValue(), company.getType(), company.getStatus(),
+            company.getManagedHubId(), company.getAddress().getValue());
     }
 
     @Override
     @Transactional
-    public UpdateCompanyResponse updateCompany(UUID id, UpdateCompanyRequestDto requestDto) {
-        // todo : 유저의 아이디 정보를 받아와서 본인 회사인지 체크하는 로직 필요
-        Company company = getOrElseThrow(id);
+    public UpdateCompanyResponse updateCompany(UpdateCompanyRequestDto requestDto) {
+        Passport passport = getPassport(requestDto.passportRequest());
+
+        if (passport.getRole().equals(AuthRole.DELIVERY_MANAGER.name())) {
+            throw new CompanyException(CompanyExceptionCode.COMPANY_ACCESS_DENIED);
+        }
+
+        Company company = getOrElseThrow(requestDto.companyId());
+
+        validHubManagerHubAndCompanyManager(passport, company);
+
         company.update(requestDto.companyName(), requestDto.companyAddress());
         return UpdateCompanyResponse.of(company.getId());
     }
 
     @Override
     @Transactional
-    public void deleteCompany(UUID id) {
-        // todo : 유저의 아이디 정보를 받아와서 본인 회사인지 체크하는 로직 필요
+    public void deleteCompany(UUID id, HttpServletRequest passportRequest) {
+        Passport passport = getPassport(passportRequest);
+
+        if (!passport.getRole().equals(AuthRole.MASTER.name()) && !passport.getRole()
+            .equals(AuthRole.HUB_MANAGER.name())) {
+            throw new CompanyException(CompanyExceptionCode.COMPANY_ACCESS_DENIED);
+        }
+
         Company company = getOrElseThrow(id);
+
+        validHubManagerHub(passport, company.getManagedHubId());
+
         companyRepository.delete(company);
     }
 
@@ -86,8 +122,20 @@ public class CompanyServiceImpl implements CompanyService {
     @Transactional
     public UpdateCompanyHubResponse updateCompanyHub(UUID id,
         UpdateCompanyHubRequestDto requestDto) {
-        // todo : 유저의 아이디 정보를 받아와서 본인 회사인지 체크하는 로직 필요
+
+        Passport passport = getPassport(requestDto.passportRequest());
+
+        if (!passport.getRole().equals(AuthRole.MASTER.name()) && !passport.getRole()
+            .equals(AuthRole.HUB_MANAGER.name())) {
+            throw new CompanyException(CompanyExceptionCode.COMPANY_ACCESS_DENIED);
+        }
+
         Company company = getOrElseThrow(id);
+        GetHubInfo hubInfo = hubServiceClient.getHubInfo(requestDto.managedHubId());
+        if (hubInfo == null) {
+            throw new CompanyException(CompanyExceptionCode.COMPANY_HUB_NOT_FOUND);
+        }
+
         company.updateHub(requestDto.managedHubId());
         return UpdateCompanyHubResponse.of(company.getId());
     }
@@ -95,7 +143,14 @@ public class CompanyServiceImpl implements CompanyService {
     @Override
     @Transactional
     public UpdateCompanyTypeResponse updateCompanyType(UpdateCompanyTypeRequestDto requestDto) {
+        Passport passport = getPassport(requestDto.passportRequest());
+
+        if (passport.getRole().equals(AuthRole.DELIVERY_MANAGER.name())) {
+            throw new CompanyException(CompanyExceptionCode.COMPANY_ACCESS_DENIED);
+        }
+
         Company company = getOrElseThrow(requestDto.companyId());
+        validHubManagerHubAndCompanyManager(passport, company);
         company.updateCompanyType(requestDto.companyType());
         return UpdateCompanyTypeResponse.of(company.getId());
     }
@@ -103,9 +158,44 @@ public class CompanyServiceImpl implements CompanyService {
     @Override
     @Transactional
     public UpdateCompanyUserResponse updateCompanyUser(UpdateCompanyUserRequestDto requestDto) {
+        Passport passport = getPassport(requestDto.passportRequest());
+        notMasterAndNotHubManagerValid(passport);
         Company company = getOrElseThrow(requestDto.companyId());
+        validHubManagerHubAndCompanyManager(passport, company);
         company.updateCompanyUser(requestDto.userId());
         return UpdateCompanyUserResponse.of(company.getId());
+    }
+
+    private Passport getPassport(HttpServletRequest passportRequest) {
+        return passportUtil.getPassportByHttpServletRequest(passportRequest);
+    }
+
+    private void notMasterAndNotHubManagerValid(Passport passport) {
+        if (passport.getRole().equals(AuthRole.COMPANY_MANAGER.name()) || passport.getRole()
+            .equals(AuthRole.DELIVERY_MANAGER.name())) {
+            throw new CompanyException(CompanyExceptionCode.COMPANY_ACCESS_DENIED);
+        }
+    }
+
+    private void validHubManagerHubAndCompanyManager(Passport passport, Company company) {
+        validHubManagerHub(passport, company.getManagedHubId());
+        if (passport.getRole().equals(AuthRole.COMPANY_MANAGER.name())) {
+            if (!company.getUserId().equals(passport.getUserId())) {
+                throw new CompanyException(CompanyExceptionCode.COMPANY_ACCESS_DENIED);
+            }
+        }
+    }
+
+    private void validHubManagerHub(Passport passport, UUID company) {
+        if (passport.getRole().equals(AuthRole.HUB_MANAGER.name())) {
+            HubManagerBooleanRequest hubManagerBooleanRequest = HubManagerBooleanRequest.of(
+                passport.getUserId(), company);
+            GetHubManagerBooleanResponse hubManagerBoolean = hubServiceClient.getHubManagerBoolean(
+                hubManagerBooleanRequest);
+            if (Boolean.FALSE.equals(hubManagerBoolean.isExist())) {
+                throw new CompanyException(CompanyExceptionCode.COMPANY_ACCESS_DENIED);
+            }
+        }
     }
 
     private Company getOrElseThrow(UUID id) {
